requirements.txt
langchain==0.3.8
langchain_google_genai==2.0.5
Pillow==11.0.0
python-dotenv==1.0.1
streamlit==1.36.0
streamlit_chat==0.1.1
transformers==4.46.3
